pallets/dpos/src/tests.rs
use crate::{mock::*, *};
use frame_support::{assert_noop, assert_ok, traits::{Currency, OnInitialize, OnFinalize}};
use frame_system::pallet_prelude::BlockNumberFor;
use rand::seq::SliceRandom;
use rand::thread_rng;
use rand::Rng;
use env_logger;
use sp_runtime::traits::Zero;


// function to advance blocks so you can run on_initialize logic or something that uses block numbers
// added setting the author dynamically based on the current validators
pub fn next_block() {
    // get list of current validators
    let current_validators = CurrentValidators::<Test>::get().into_inner();
    // If there are current validators, choose one randomly as the author
    if !current_validators.is_empty() {
        let mut rng = thread_rng();
        if let Some(&author) = current_validators.choose(&mut rng) {
            mock::Author::set(author);
        }
    }
    System::set_block_number(System::block_number() + 1);
    System::on_initialize(System::block_number());
    Dpos::on_initialize(System::block_number());
}

pub fn run_to_block(n: BlockNumberFor<Test>) {
    while System::block_number() < n {
        if System::block_number() > 1 {
            Dpos::on_finalize(System::block_number());
            System::on_finalize(System::block_number());

        }
        next_block();
    }
}

// helper function to print the total delegated stake for each elected validator
fn print_total_delegated_stake_for_elected<T: Config>() {
    log::info!("Total delegated stake for each elected validator:");
    let current_validators = CurrentValidators::<T>::get();
    for validator in current_validators.into_inner() {
        let stake = ValidatorStakes::<T>::get(&validator);
        log::info!("Validator {:?}: total delegated stake = {:?}", validator, stake);
    }
}

// ----- tests -----

#[test]
fn it_works_for_default_value() {
	new_test_ext().execute_with(|| {
		// Go past genesis block so events get deposited - block number to 1 to emit events
		run_to_block(1);

		// Fund an account with sufficient balance
		Balances::make_free_balance_be(&55, 1_000);

		// Test registering a validator
		assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

		// Verify storage updates
		assert!(PotentialValidators::<Test>::contains_key(&55));
		assert_eq!(PotentialValidators::<Test>::get(&55), Some(500));

		// Verify event emission
		System::assert_last_event(Event::ValidatorRegistered { validator: 55, amount: 500 }.into());
	
	});
} 

// ------ register validator tests -------

// test for registering a validator
#[test]
fn register_validator() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);

        // Test registering a validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Verify storage updates
        assert!(PotentialValidators::<Test>::contains_key(&55));
        assert_eq!(PotentialValidators::<Test>::get(&55), Some(500));

        // Verify event emission
        System::assert_last_event(Event::ValidatorRegistered { validator: 55, amount: 500 }.into());
    });
}

// register multiple validators from different accounts
#[test]
fn register_multiple_validators() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund accounts with sufficient balance
        for i in 11..=15 {
            Balances::make_free_balance_be(&i, 1_000);
        }

        // Register validators
        for i in 11..=15 {
            assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(i), 500));
        }

        // Verify storage updates
        for i in 11..=15 {
            assert!(PotentialValidators::<Test>::contains_key(&i));
            assert_eq!(PotentialValidators::<Test>::get(&i), Some(500));
        }

        // Verify event emission
        System::assert_last_event(Event::ValidatorRegistered { validator: 15, amount: 500 }.into());
    });
}

// test validator registering with insufficient balance
#[test]
fn register_validator_with_insufficient_balance() {
	new_test_ext().execute_with(|| {
		run_to_block(1);

		Balances::make_free_balance_be(&55, 499);

		// Test registering a validator with insufficient balance
		assert_noop!(
			Dpos::register_validator(RuntimeOrigin::signed(55), 500),
			Error::<Test>::InsufficientBalance
		);
	});
}

// test validator trying to register twice
#[test]
fn register_validator_twice() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Test registering a validator twice
        assert_noop!(
            Dpos::register_validator(RuntimeOrigin::signed(55), 500),
            Error::<Test>::AlreadyRegistered
        );
    });
}


// ----- unregister validator tests -------

// test for unregistering a validator
#[test]
fn unregister_validator() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        // Set up initial balances
        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);
        Balances::make_free_balance_be(&77, 1_000);

        // Register validator and delegate
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(77), 55, 200));

        // Unregister the validator
        assert_ok!(Dpos::unregister_validator(RuntimeOrigin::signed(55)));

        // Verify storage updates
        assert!(!PotentialValidators::<Test>::contains_key(&55));
        assert!(!ValidatorStakes::<Test>::contains_key(&55));
        assert!(!Delegators::<Test>::contains_key(&66));
        assert!(!Delegators::<Test>::contains_key(&77));

        // Verify event emission
        System::assert_last_event(Event::ValidatorDeregistered {
            validator: 55,
        }
        .into());
    });
}

// attempts to unregister an unregistered validator
#[test]
fn unregister_unregistered_validator() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);

        // Attempt to unregister an unregistered validator
        assert_noop!(
            Dpos::unregister_validator(RuntimeOrigin::signed(55)),
            Error::<Test>::ValidatorNotFound
        );
    });
}

// attempts to unregister a validator with no delegators
#[test]
fn unregister_validator_with_no_delegators() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);

        // Register validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Attempt to unregister a validator with no delegators
        assert_ok!(Dpos::unregister_validator(RuntimeOrigin::signed(55)));

        // Verify storage updates
        assert!(!PotentialValidators::<Test>::contains_key(&55));
        assert!(!ValidatorStakes::<Test>::contains_key(&55));
    });
}

// register a validator and then immediately deregister
#[test]
fn register_and_deregister_validator() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);

        // Register validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Deregister the validator
        assert_ok!(Dpos::unregister_validator(RuntimeOrigin::signed(55)));

        // Verify storage updates
        assert!(!PotentialValidators::<Test>::contains_key(&55));
        assert!(!ValidatorStakes::<Test>::contains_key(&55));
    });
}

// ------ delegate tests -------

// test for delegation to a validator
#[test]
fn delegate_to_validator() {
	new_test_ext().execute_with(|| {
		run_to_block(1);

		Balances::make_free_balance_be(&55, 1_000);
		Balances::make_free_balance_be(&66, 1_000);

		assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

		assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

		// Verify storage updates
		assert!(Delegators::<Test>::contains_key(&66));
		let delegation = Delegators::<Test>::get(&66).unwrap();
		assert_eq!(delegation.validator, 55);
		assert_eq!(delegation.amount, 300);

		// Verify event emission
		System::assert_last_event(Event::Delegated { delegator: 66, validator: 55, amount: 300 }.into());
	});
}

// test to check if a delegator can delegate to a validator that is not registered
#[test]
fn delegate_to_unregistered_validator() {
	new_test_ext().execute_with(|| {
		run_to_block(1);

		Balances::make_free_balance_be(&55, 1_000);
		Balances::make_free_balance_be(&66, 1_000);

		assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

		// Test delegating to an unregistered validator
		assert_noop!(
			Dpos::delegate(RuntimeOrigin::signed(66), 66, 300),
			Error::<Test>::ValidatorNotFound
		);
	});
}

// test to check if a delegator can delegate more than their balance
#[test]
fn delegate_more_than_balance() {
	new_test_ext().execute_with(|| {
		run_to_block(1);

		Balances::make_free_balance_be(&55, 1_000);
		Balances::make_free_balance_be(&66, 1_000);

		assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

		// Test delegating more than the delegator's balance
		assert_noop!(
			Dpos::delegate(RuntimeOrigin::signed(66), 55, 1_001),
			Error::<Test>::InsufficientBalance
		);
	});
}

// test delegator trying to delegate to more than one validator
#[test]
fn delegate_to_multiple_validators() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);
        Balances::make_free_balance_be(&77, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(66), 500));

        // Delegate to a validator
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(77), 55, 300));

        // Attempt to delegate to another validator should fail
        assert_noop!(
            Dpos::delegate(RuntimeOrigin::signed(77), 66, 300),
            Error::<Test>::AlreadyDelegated
        );

        // Verify storage updates
        assert!(Delegators::<Test>::contains_key(&77));
        let delegation = Delegators::<Test>::get(&77).unwrap();
        assert_eq!(delegation.validator, 55);
        assert_eq!(delegation.amount, 300);

        // Verify event emission
        System::assert_last_event(Event::Delegated { delegator: 77, validator: 55, amount: 300 }.into());
    });
}

// verify storage update after a successful delegation
#[test]
fn delegate_storage_update() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        // Verify storage updates
        assert!(Delegators::<Test>::contains_key(&66));
        let delegation = Delegators::<Test>::get(&66).unwrap();
        assert_eq!(delegation.validator, 55);
        assert_eq!(delegation.amount, 300);
    });
}

// try to delegate zero
#[test]
fn delegate_zero_balance() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Test delegating zero balance
        assert_noop!(
            Dpos::delegate(RuntimeOrigin::signed(66), 55, 0),
            Error::<Test>::InvalidAmount
        );
    });
}

// Test Delegation After Unregistering Validator
#[test]
fn delegate_after_unregistering_validator() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        // Register and then unregister a validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::unregister_validator(RuntimeOrigin::signed(55)));

        // Attempt to delegate to an unregistered validator
        assert_noop!(
            Dpos::delegate(RuntimeOrigin::signed(66), 55, 300),
            Error::<Test>::ValidatorNotFound
        );
    });
}

// Ensure delegating an amount exactly equal to the available balance works correctly
#[test]
fn delegate_with_exact_balance() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund accounts with sufficient balance considering a small transaction fee (e.g., 1 unit)
        let balance = 1_000;
        let fee = 1;
        let delegation_amount = balance - fee;

        Balances::make_free_balance_be(&55, balance);
        Balances::make_free_balance_be(&66, balance);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Test delegating the exact available balance minus the fee
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, delegation_amount));

        // Verify storage updates
        assert!(Delegators::<Test>::contains_key(&66));
        let delegation = Delegators::<Test>::get(&66).unwrap();
        assert_eq!(delegation.validator, 55);
        assert_eq!(delegation.amount, delegation_amount);

        // Verify event emission
        System::assert_last_event(Event::Delegated { delegator: 66, validator: 55, amount: delegation_amount }.into());
    });
}


// ------- undelegatiuon tests -------

// successful partial undelegetion
#[test]
fn partial_undelegate() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        // Test partial undelegation
        assert_ok!(Dpos::undelegate(RuntimeOrigin::signed(66), 100));

        // Verify storage updates
        assert!(Delegators::<Test>::contains_key(&66));
        let delegation = Delegators::<Test>::get(&66).unwrap();
        assert_eq!(delegation.validator, 55);
        assert_eq!(delegation.amount, 200);

        // Verify event emission
        System::assert_last_event(Event::Undelegated { delegator: 66, validator: 55, amount: 100 }.into());
    });
}

// test for undelegation with no delegation found
#[test]
fn undelegation_no_delegation_found() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund accounts with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        // Register validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));

        // Attempt to undelegate without delegation
        assert_noop!(
            Dpos::undelegate(RuntimeOrigin::signed(66), 200),
            Error::<Test>::NoDelegationFound
        );
    });
}

// test for undelegation with insufficient balance
#[test]
fn undelegation_insufficient_balance() {
    new_test_ext().execute_with(|| {
        run_to_block(1);

        // Fund accounts with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        // Register validator and delegate
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        // Attempt to undelegate more than delegated amount
        assert_noop!(
            Dpos::undelegate(RuntimeOrigin::signed(66), 400),
            Error::<Test>::InsufficientBalance
        );
    });
}

// test complete undelegation -- atm not working bc im not removing the delegator from the delegators storage
#[test]
fn complete_undelegation() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));
        
        // Test complete undelegation
        assert_ok!(Dpos::undelegate(RuntimeOrigin::signed(66), 300));

        // Verify storage updates
        assert!(!Delegators::<Test>::contains_key(&66));

        let validator_stake = ValidatorStakes::<Test>::get(&55);
        assert_eq!(validator_stake, 500);

        // Verify event emission
        System::assert_last_event(Event::Undelegated {
            delegator: 66,
            validator: 55,
            amount: 300,
        }
        .into());
    });
}

// delegator can successfully re-delegate after undelegating
#[test]
fn redelegate_after_undelegation() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        // Test undelegation
        assert_ok!(Dpos::undelegate(RuntimeOrigin::signed(66), 300));

        // Test re-delegation
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        // Verify storage updates
        assert!(Delegators::<Test>::contains_key(&66));
        let delegation = Delegators::<Test>::get(&66).unwrap();
        assert_eq!(delegation.validator, 55);
        assert_eq!(delegation.amount, 300);

        // Verify event emission
        System::assert_last_event(Event::Delegated { delegator: 66, validator: 55, amount: 300 }.into());
    });
}



// ------ update validator set tests -------

// test for validator set update based on total stake
// add 10 validators and for each validator 3 delegators,select the top 5 validators 
// then stake more and update the validator set again and check if the top 5 validators upd
// added dynamic testing and removed hardcoding - tested for a large number of validators
// prints the logs from the functions in order to check that everything is working as expected
// took out the assert_eq! for the validators as it was failing due to the randomness of the stake

#[test]
fn update_validator_set() {
    new_test_ext().execute_with(|| {

		let _ = env_logger::builder().is_test(true).try_init();

        run_to_block(99);

        let pool: u64 = 20;

        // Fund accounts with sufficient balance
        for i in 11..=pool {
            Balances::make_free_balance_be(&i, 1_000);
        }

        // Register validators
        for i in 11..=pool {
            assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(i), 100));
        }

        // Delegate to validators
        for i in 11..=pool {
            for j in 11..=15 {
                let num = rand::thread_rng().gen_range(9..50); // Generate random number in the range
                let delegator = j * pool + i; // create unique delegator ids
                Balances::make_free_balance_be(&delegator, 10_000);
                let amount = (i * num).into(); // delegate different amounts based on `i`
                assert_ok!(Dpos::delegate(RuntimeOrigin::signed(delegator), i, amount));
            }
        }

        // Advance to the next block to trigger validator update
        run_to_block(100);

        // Verify storage updates
        let current_validators = CurrentValidators::<Test>::get();
        // log::info!("Current validators: {:?}", current_validators);
        let max_validators = <Test as crate::Config>::MaxValidators::get() as usize;
        assert_eq!(current_validators.len(), max_validators);
		//let expected_validators: Vec<u64> = (6..=15).rev().collect();
        let max_validators_u64 = max_validators as u64;
        //let expected_validators: Vec<u64> = ((pool - max_validators_u64 + 1)..=pool).rev().collect();
        //assert_eq!(current_validators.into_inner(), expected_validators);

        // Verify event emission
		let events = System::events();
        assert!(events.iter().any(|record| {
            matches!(record.event, RuntimeEvent::Dpos(Event::ValidatorsUpdated))
        }));

		// Print total stake for each validator
        print_total_delegated_stake_for_elected::<Test>();

		let poor: u64 = (pool - max_validators_u64)/2;
        // Second round: Add more stake to the lower ranked validators
        for i in 11..=poor {
            for j in 11..=13 {
                let num = rand::thread_rng().gen_range(9..50); // Generate random number in the range 
                let delegator = j * pool + i; // same delegators
                assert_ok!(Dpos::delegate(RuntimeOrigin::signed(delegator), i, (i * num).into()));
            }
        }

		// advance to the next block to trigger validator update
		run_to_block(200);

		// Verify storage updates
		let current_validators = CurrentValidators::<Test>::get();
		// log::info!("Current validators: {:?}", current_validators);
		assert_eq!(current_validators.len(), max_validators);

		// Verify event emission
		let events = System::events();
        assert!(events.iter().any(|record| {
            matches!(record.event, RuntimeEvent::Dpos(Event::ValidatorsUpdated))
        }));

        print_total_delegated_stake_for_elected::<Test>();

    });
}

// test fair epoch reward distribution for validators and delegators
#[test]
fn delegators_receive_rewards_only_after_next_epoch() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000); // Validator
        Balances::make_free_balance_be(&66, 1_000); // Delegator 1
        Balances::make_free_balance_be(&77, 1_000); // Delegator 2
        Balances::make_free_balance_be(&88, 1_000); // Delegator 3

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300)); // delegator 1 delegates beginning of epoch

        let initial_b_validator = Balances::free_balance(&55); 
        let initial_b_delegator1 = Balances::free_balance(&66); // balance delegator 1 after staking

        run_to_block(150);  // Advance to half of the epoch duration

        // delegato 2 delegates in the middle of the epoch
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(77), 55, 200));
        let initial_b_delegator2 = Balances::free_balance(&77); // balance after delegation

        run_to_block(190); // go to just before end of epoch 

        let final_b_delegator2 = Balances::free_balance(&77);
        assert_eq!(final_b_delegator2, initial_b_delegator2);  // still in current epoch, Delegator 2 should not receive rewards until the next epoch
       
        run_to_block(200); // go to next epoch

        // Verify rewards are distributed correctly
        let final_b_delegator1 = Balances::free_balance(&66);
        let final_b_validator = Balances::free_balance(&55);

        assert!(final_b_delegator1 > initial_b_delegator1); // Delegator 1 should receive rewards
        assert!(final_b_validator > initial_b_validator); // Validator should receive rewards

        run_to_block(250);

        // delegato 3 delegates in the middle of the epoch
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(88), 55, 200));
        let initial_b_delegator3 = Balances::free_balance(&88); // balance after delegation

        run_to_block(290); // go to just before end of epoch

        let final_b_delegator3 = Balances::free_balance(&88);
        assert_eq!(final_b_delegator3, initial_b_delegator3);  // still in current epoch, Delegator 2 should not receive rewards until the next epoch


        run_to_block(300); // go to next epoch

        let final_b_delegator2_epoch_2 = Balances::free_balance(&77);
        assert!(final_b_delegator2_epoch_2 > initial_b_delegator2); // now he should receive rewards

        run_to_block(400); 

        let final_b_delegator3_epoch_3 = Balances::free_balance(&88);
        assert!(final_b_delegator3_epoch_3 > initial_b_delegator3); // now he should receive rewards

    });
}

// ----- slash tests -------
#[test]
fn slash_validator_successfully() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        // Register a validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        // Slash the validator
        assert_ok!(Dpos::test_slash_validator(RuntimeOrigin::root(), 55));

        // Verify storage updates
        assert!(!PotentialValidators::<Test>::contains_key(&55));
        assert_eq!(ValidatorStakes::<Test>::get(&55), Zero::zero());

        // Verify event emission
        System::assert_last_event(Event::ValidatorSlashed { validator: 55, amount: 800 }.into());
    });
}

// test for slashing a non registered validator
#[test]
fn slash_non_registered_validator() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        // Attempt to slash a non-registered validator
        assert_noop!(
            Dpos::test_slash_validator(RuntimeOrigin::root(), 55),
            Error::<Test>::ValidatorNotFound
        );
    });
}

// test for slashing a validator with zero stake
#[test]
fn slash_validator_with_zero_stake() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);

        // Register and then unregister a validator
        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::unregister_validator(RuntimeOrigin::signed(55)));

        // Attempt to slash a validator with zero stake
        assert_noop!(
            Dpos::test_slash_validator(RuntimeOrigin::root(), 55),
            Error::<Test>::ValidatorNotFound
        );
    });
}

// Ensure slashing attempts fail for non-validator accounts
#[test]
fn slash_non_validator() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        // Fund an account with sufficient balance
        Balances::make_free_balance_be(&55, 1_000);

        // Attempt to slash a non-validator account
        assert_noop!(
            Dpos::test_slash_validator(RuntimeOrigin::root(), 55),
            Error::<Test>::ValidatorNotFound
        );
    });
}

// test system handling multiple epoch transitions correctly, including reward distribution
#[test]
fn multiple_epoch_transitions() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);
        Balances::make_free_balance_be(&77, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        let initial_b_validator = Balances::free_balance(&55);
        let initial_b_delegator = Balances::free_balance(&66);

        // Simulate multiple epoch transitions
        for _ in 0..3 {
            run_to_block(System::block_number() + 100);
        }

        // Verify rewards are distributed correctly
        let final_b_validator = Balances::free_balance(&55);
        let final_b_delegator = Balances::free_balance(&66);

        assert!(final_b_delegator > initial_b_delegator); // Delegator should receive rewards
        assert!(final_b_validator > initial_b_validator); // Validator should receive rewards
    });
}


// Ensure removing delegators and validators mid-epoch works correctly and rewards are adjusted appropriately.
// as i wrote in the readme this should be handled more appropriately but for now testing that 
// they are removed without receiving rewards 
#[test]
fn remove_delegator_mid_epoch() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        run_to_block(50); // Mid-epoch

        // Remove the delegator
        assert_ok!(Dpos::undelegate(RuntimeOrigin::signed(66), 300));

        // Verify storage updates
        assert!(!Delegators::<Test>::contains_key(&66));
        let validator_stake = ValidatorStakes::<Test>::get(&55);
        assert_eq!(validator_stake, 500);

        // Verify event emission
        System::assert_last_event(Event::Undelegated {
            delegator: 66,
            validator: 55,
            amount: 300,
        }
        .into());

        run_to_block(100); // Complete epoch

        // Verify no rewards for the removed delegator
        let final_b_delegator = Balances::free_balance(&66);
        assert_eq!(final_b_delegator, 1_000); // No change from initial balance
    });
}

#[test]
fn remove_validator_mid_epoch() {
    new_test_ext().execute_with(|| {
        let _ = env_logger::builder().is_test(true).try_init();
        run_to_block(1);

        Balances::make_free_balance_be(&55, 1_000);
        Balances::make_free_balance_be(&66, 1_000);

        assert_ok!(Dpos::register_validator(RuntimeOrigin::signed(55), 500));
        assert_ok!(Dpos::delegate(RuntimeOrigin::signed(66), 55, 300));

        run_to_block(50); // Mid-epoch

        // Remove the validator
        assert_ok!(Dpos::unregister_validator(RuntimeOrigin::signed(55)));

        // Verify storage updates
        assert!(!PotentialValidators::<Test>::contains_key(&55));
        assert!(!ValidatorStakes::<Test>::contains_key(&55));
        assert!(!Delegators::<Test>::contains_key(&66));

        // Verify event emission
        System::assert_last_event(Event::ValidatorDeregistered {
            validator: 55,
        }
        .into());

        run_to_block(100); // Complete epoch

        // Verify no rewards for the removed validator and delegator
        let final_b_validator = Balances::free_balance(&55);
        let final_b_delegator = Balances::free_balance(&66);

        assert_eq!(final_b_validator, 1_000); // No change from initial balance
        assert_eq!(final_b_delegator, 1_000); // No change from initial balance
    });
}



